% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorUtils.r
\docType{package}
\name{anchorUtils}
\alias{anchorUtils}
\title{anchorUtils: In-house utility functions for Anchorlytics Consulting}
\description{
This package is a dump-bin of miscellaneous helper functions accumulated
over the course of several projects.
}
