% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{drop_nzv}
\alias{drop_nzv}
\title{Near-Zero Variance filter}
\usage{
drop_nzv(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{other options passed through to \code{\link[caret:nearZeroVar]{caret::nearZeroVar()}}}
}
\value{
data frame, potentially with fewer columns
}
\description{
Removes numeric variables with very little variability.
This can, for instance, aid stability of inverting covariance matrices for
factor analysis.
}
\section{TODO}{
\itemize{
\item remove dependency on caret
}
}

\examples{
drop_nzv(mtcars, freqCut = 1.4)

}
\seealso{
Other data munging: 
\code{\link{dummy_code}()},
\code{\link{ord_collapse}()}
}
\concept{data munging}
