% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Show only duplicated rows}
\usage{
duplicates(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification of columns}
}
\value{
grouped tibble of duplicate rows
}
\description{
Whereas \code{\link[dplyr:distinct]{dplyr::distinct()}} removes duplicates, this function
removes unique rows and only shows rows whose values in the specified
columns are duplicates.
}
\examples{
duplicates(data.frame(x = 1:5, y = c(1, 2, 1, 4, 5)), dplyr::everything())

}
