% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelia.R
\name{get_ords}
\alias{get_ords}
\title{List all ordered factors in a data frame}
\usage{
get_ords(data)
}
\arguments{
\item{data}{data frame}
}
\value{
character vector of column names
}
\description{
List all ordered factors in a data frame
}
\examples{
get_ords(data.frame(
  x = factor(LETTERS[1:5]),
  y = factor(LETTERS[6:10], ordered = TRUE)
))

}
\seealso{
Other Amelia imputation: 
\code{\link{get_noms}()},
\code{\link{run_amelia}()}
}
\concept{Amelia imputation}
