% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{is_identified}
\alias{is_identified}
\title{Check if fitted mplusObject has converged}
\usage{
is_identified(.obj, param = "se")
}
\arguments{
\item{.obj}{mplusObject fitted by MplusAutomation::mplusModeler()}

\item{param}{character name of field to look for in parameter list.
By default, look for existence of standard errors.}
}
\value{
logical
}
\description{
This macro uses the existence of estimated standard errors as a proxy
for determining if the model was identified.
}
\seealso{
Other Mplus helpers: \code{\link{mp_cat}},
  \code{\link{mp_run}}, \code{\link{mp_wordwrap}},
  \code{\link{parse_tech4}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
