% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\docType{data}
\name{mp_abbrev}
\alias{mp_abbrev}
\title{Abbreviations for common phrases in Mplus messages}
\format{An object of class \code{character} of length 4.}
\usage{
mp_abbrev
}
\description{
Abbreviations for common phrases in Mplus messages
}
\examples{
stringr::str_replace_all(c(
  "cov matrix (psi) is not positive definite. Variable x."
  ), mp_abbrev)
}
\seealso{
Other Mplus helpers: \code{\link{is_identified}},
  \code{\link{mp_cat}}, \code{\link{mp_messages}},
  \code{\link{mp_run}}, \code{\link{mp_wordwrap}},
  \code{\link{parse_tech4}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
\keyword{datasets}
