% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_cat}
\alias{mp_cat}
\title{Specify categorical variables to Mplus}
\usage{
mp_cat(.data, ...)
}
\arguments{
\item{.data}{data frame with proper column types}

\item{...}{additional options passed to [mp_wordwrap()]}
}
\value{
a single string that can be passed to
  [MplusAutomation::mplusObject()] in the `VARIABLE` argument
}
\description{
This creates an Mplus input syntax line indicating which
variables are categorical (ordinal or binary/dichotomous).
A variable is deemed dichotomous if it has at most two unique non-NA values,
regardless of its type.
}
\examples{
cat(mp_cat(data.frame(
  Number_of_Cylinders = ordered(mtcars$cyl),
  Miles_per_Gallon = mtcars$mpg,
  Engine_V_shaped_or_Straight = ordered(mtcars$vs),
  Transmission_Automatic_or_Manual = ordered(mtcars$am))))
}
\seealso{
Other Mplus helpers: 
\code{\link{is_identified}()},
\code{\link{mp_abbrev}},
\code{\link{mp_messages}()},
\code{\link{mp_run}()},
\code{\link{mp_wordwrap}()},
\code{\link{parse_tech4}()}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
