% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_messages}
\alias{mp_messages}
\title{Format Mplus warning/error messages}
\usage{
mp_messages(.obj, what = c("errors", "warnings"))
}
\arguments{
\item{.obj}{mplusObject fitted by MplusAutomation::mplusModeler()}

\item{what}{character vector of fields to pull from .obj}
}
\value{
character vector of messages
}
\description{
Pull all messages from the Mplus output object,
concatenate multi-line messages,
and convert case so messages are not all-caps.
}
\seealso{
Other Mplus helpers: 
\code{\link{is_identified}()},
\code{\link{mp_abbrev}},
\code{\link{mp_cat}()},
\code{\link{mp_run}()},
\code{\link{mp_wordwrap}()},
\code{\link{parse_tech4}()}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
