% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_ordinal}
\alias{mp_ordinal}
\title{Specify ordinal variables to Mplus}
\usage{
mp_ordinal(.data, ...)
}
\arguments{
\item{.data}{data frame with proper column types}

\item{...}{additional options passed to [mp_wordwrap()]}
}
\value{
a single string that can be passed to
  [MplusAutomation::mplusObject()] in the `VARIABLE` argument
}
\description{
This creates an Mplus input syntax line indicating which
variables are ordered categorical.
}
\examples{
cat(mp_ordinal(data.frame(
  Number_of_Cylinders = ordered(mtcars$cyl),
  Miles_per_Gallon = mtcars$mpg,
  Engine_V_shaped_or_Straight = ordered(mtcars$vs),
  Transmission_Automatic_or_Manual = ordered(mtcars$am))))
}
\seealso{
Other Mplus helpers: \code{\link{mp_wordwrap}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
