% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_run}
\alias{mp_run}
\title{Run Mplus and save full output}
\usage{
mp_run(.obj, name = "mp", run = 1, ...)
}
\arguments{
\item{.obj}{mplusObject as created by MplusAutomation::mplusObject()}

\item{name}{basename for generated Mplus files (no extension)}

\item{run}{set to 0 for dry-run (passed to MplusAutomation::mplusModeler())}

\item{...}{additional options passed to MplusAutomation::mplusModeler()}
}
\value{
an mplusObject with a $results section
}
\description{
MplusAutomation::mplusModeler() runs Mplus and augments the
mplusOjbect with a $results element.
This wrapper saves the full text output from Mplus in an element named
$stdout within that $results element.
}
\details{
This function has side-effects, just like mplusModeler(), in that
it leaves Mplus input, output, and savedata files in the current directory.
}
\seealso{
Other Mplus helpers: \code{\link{is_identified}},
  \code{\link{mp_abbrev}}, \code{\link{mp_cat}},
  \code{\link{mp_messages}}, \code{\link{mp_wordwrap}},
  \code{\link{parse_tech4}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
