% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_wordwrap}
\alias{mp_wordwrap}
\title{Word wrap to accommodate Mplus line length limitations.}
\usage{
mp_wordwrap(.list, width = 80, exdent = 4, ...)
}
\arguments{
\item{.list}{a character vector}

\item{width}{passed to [base::strwrap()]}

\item{exdent}{passed to [base::strwrap()]}

\item{...}{passed to [base::strwrap()]}
}
\value{
a single string, potentially with newlines
}
\description{
Mplus input syntax is limited to 90 characters on a line.
This function takes a character vector, concatenates it with spaces,
and word-wraps the result, returning a single string.
}
\examples{
cat(mp_wordwrap(rownames(mtcars)))
}
\seealso{
Other Mplus helpers: \code{\link{is_identified}},
  \code{\link{mp_abbrev}}, \code{\link{mp_cat}},
  \code{\link{mp_messages}}, \code{\link{mp_run}},
  \code{\link{parse_tech4}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
