% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{ord_collapse}
\alias{ord_collapse}
\title{Compress integer ordinal factors with too many levels}
\usage{
ord_collapse(.data, maxlev = 10, newlev = maxlev)
}
\arguments{
\item{.data}{data frame with ordinal factors coded as integers}

\item{maxlev}{apply level collapsing only if \code{nlevels} is greater than this}

\item{newlev}{if collapsing levels, how many levels should the factor have.
Default is \code{maxlev}.}
}
\value{
data frame with same columns
}
\description{
Polychoric correlation routines often limit the number of levels an
ordinal (ordered categorical) factor can have.
This function is like \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}
but preserves the ordering of levels.
Modified factors are releveled to 1:n.
}
\examples{
library(dplyr)
mtcars \%>\%
  mutate(across(c("cyl", "vs", "am", "gear", "carb", "disp"), as.ordered)) \%>\%
  ord_collapse() \%>\%
  summarise_all(nlevels)

}
\seealso{
Other data munging: 
\code{\link{drop_nzv}()},
\code{\link{dummy_code}()}
}
\concept{data munging}
