% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{parse_tech4}
\alias{parse_tech4}
\title{Parse Mplus TECH4 output}
\usage{
parse_tech4(.text)
}
\arguments{
\item{.text}{contents of an Mplus output file, as read by readLines()}
}
\value{
a list with one entry for each group.
  Each entry has members 'cov' and 'cor',
  each of which has members 'est', 'se', 'ratio', and 'p',
  each of which is a matrix.
}
\description{
Mplus' `TECH4` output contains first- and second-order estimated **moments**
for the latent factors.
Unfortunately, the `MplusAutomation` library is still buggy
in parsing this section; hence this function.
}
\details{
For **single-group** models, the TECH4 section has one subsection,
containing the following moments estimated for the latent factors:

1. **Means**
2. **Covariance** matrix
3. **Correlation** matrix

If **standard errors** are available, a second subsection is added with
the following for **each** of the means, covariance, and correlation:

1. **Standard error**
2. **Ratio** of SE/estimate
3. **P-value** on estimate differing from zero

Hence, there are 9 items within this second subsection.

For **multi-group** models, each group has its own subsection.
If standard errors are available, there are **twice** as many subsections as
there are groups: first all the moments for each of the groups, then all the
standard errors for each of the groups.

TODO: if no groups, returns a single group named "X"
TODO: extract vector of means
}
\seealso{
Other Mplus helpers: 
\code{\link{is_identified}()},
\code{\link{mp_abbrev}},
\code{\link{mp_cat}()},
\code{\link{mp_messages}()},
\code{\link{mp_run}()},
\code{\link{mp_wordwrap}()}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
