% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all.r
\name{read_all}
\alias{read_all}
\title{Load multiple RDS files into current environment}
\usage{
read_all(.list, ...)
}
\arguments{
\item{.list}{character vector of names}

\item{...}{path components relative to project root}
}
\value{
None, global environment is modified
}
\description{
Load multiple RDS files into current environment
}
\examples{
# Reads from "data/rds/my_mtcars.rds" and "data/rds/my_letters.rds"
# relative to project root
read_all(c("my_mtcars", "my_letters"), "data", "rds")
}
