% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelia.R
\name{run_amelia}
\alias{run_amelia}
\title{Run Amelia imputation on data with factors}
\usage{
run_amelia(x, boot.type = "none", ...)
}
\arguments{
\item{x}{matrix, data.frame, "amelia" object, or "molist" object: input data}

\item{boot.type}{"ordinary" to enable bootstrap stderr estimates}

\item{...}{other args passed to \code{\link[Amelia:amelia]{Amelia::amelia()}}}
}
\value{
"amelia" object containing imputed datasets
}
\description{
Wrapper around \code{\link[Amelia:amelia]{Amelia::amelia()}} filling in the \code{ords} and \code{noms} args.
}
\seealso{
Other Amelia imputation: 
\code{\link{get_noms}()},
\code{\link{get_ords}()}
}
\concept{Amelia imputation}
