% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tgt_time}
\alias{tgt_time}
\title{Wrapper around \code{\link[targets:tar_meta]{targets::tar_meta()}}}
\usage{
tgt_time(names = NULL, ...)
}
\arguments{
\item{names}{char vec names of targets}

\item{...}{other args passed to \code{\link[targets:tar_meta]{targets::tar_meta()}}}
}
\value{
single char string, formatted duration
}
\description{
Wrapper around \code{\link[targets:tar_meta]{targets::tar_meta()}}
}
\concept{target helpers}
