% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorUtils.R
\docType{package}
\name{anchorUtils}
\alias{anchorUtils}
\title{anchorUtils: In-house utility functions for Anchorlytics Consulting}
\description{
This package is a dump-bin of miscellaneous helper functions accumulated
over the course of several projects.
}
\section{saving/loading data}{

Functions to quickly save/load a set of R data structures.
}

\section{data munging}{

Functions to help process datasets in preparation for analysis.
}

\section{VR12 utilities}{

Functions to derive composite VR12 MCS/PCS PROM scores.
}

