% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{dummy_code}
\alias{dummy_code}
\title{Dummy-code (one-hot) factors}
\usage{
dummy_code(.data, ...)
}
\arguments{
\item{.data}{data frame or tibble}

\item{...}{One or more unquoted expressions separated by commas,
as passed to [dplyr::select()].
If omitted, all variables are considered.}
}
\value{
A tibble with selected factors converted to logical dummy variables
}
\description{
Dummy coding expands a polytomous categorical factor into several
indicator variables, each of which is logical (Boolean, TRUE/FALSE).
The one-hot contrast expands a factor with `L` levels into `L` indicator
variables.  The result is not a full-rank parameterisation,
however it avoids needing to select a reference level.
}
\details{
Ordinal (ordered categorical) and dichotomous (only 2 levels) variables are
not modified; nor are non-factors (numeric, logical, etc.).
Additional arguments are passed to [dplyr::select()] to further filter the
list of variables eligible for dummy coding.

The new indicator variables are appended to the end of the variable list,
and the original factors are removed.
}
\examples{
mutate_at(mtcars, c("cyl", "gear"), as.factor) \%>\% dummy_code()

mutate_at(mtcars, c("cyl", "gear"), as.factor) \%>\% dummy_code(-cyl)
}
\seealso{
Other data munging: \code{\link{drop_nzv}},
  \code{\link{ord_collapse}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{data munging}
