% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_time.r
\name{fmt.proc_time}
\alias{fmt.proc_time}
\title{Pretty-print a proc_time object}
\usage{
fmt.proc_time(.pt)
}
\arguments{
\item{.pt}{A proc_time object, e.g., from \code{\link[base:system.time]{base::system.time()}}}
}
\value{
A formatted string.
}
\description{
Pretty-print a proc_time object
}
\section{TODO}{
\itemize{
\item replace purrr with lapply
}
}

\examples{
fmt.proc_time(system.time(rnorm(1e7)))

}
