% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.r
\name{mp_ordinal}
\alias{mp_ordinal}
\title{Specify ordinal variables to Mplus}
\usage{
mp_ordinal(.data, ...)
}
\arguments{
\item{.data}{data frame with proper column types}

\item{...}{additional options passed to [mp_wordwrap()]}
}
\value{
a single string that can be passed to
  [MplusAutomation::mplusObject()] in the `VARIABLE` argument
}
\description{
This creates an Mplus input syntax line indicating which
variables are ordered categorical.
}
\examples{
library(dplyr)
as_tibble(mtcars) \%>\%
  mutate_at(c("cyl", "vs", "am"), as.ordered) \%>\%
  rename(`Number of Cylinders` = cyl,
         `Engine V-shaped or Straight` = vs,
         `Transmission Automatic or Manual` = am) \%>\%
  mp_ordinal() \%>\%
  cat()
}
\seealso{
Other Mplus helpers: \code{\link{mp_wordwrap}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Mplus helpers}
