% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all.r
\name{read_all}
\alias{read_all}
\title{Load multiple RDS files into current environment}
\usage{
read_all(.list, rdsdir = NULL, ...)
}
\arguments{
\item{.list}{character vector of names}

\item{rdsdir}{path to directory, passed to [base::file.path()].
If not specified, [base::tempdir()].}

\item{...}{additional options passed to [base::readRDS()]}
}
\value{
None, global environment is modified
}
\description{
Load multiple RDS files into current environment
}
\examples{
# Saves/loads files "mtcars.rds" and "letters.rds" in tmpdir
my_mtcars <- mtcars; my_letters <- letters
save_all(c("my_mtcars", "my_letters"))
rm(my_mtcars, my_letters)
read_all(c("my_mtcars", "my_letters"))
}
\seealso{
Other Saving/loading data: 
\code{\link{save_all}()}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Saving/loading data}
