% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all.R
\name{read_all}
\alias{read_all}
\title{Load multiple RDS files into current environment}
\usage{
read_all(.list, rdsdir = NULL, ...)
}
\arguments{
\item{.list}{character vector of names}

\item{rdsdir}{path to directory, passed to \code{\link[base:file.path]{base::file.path()}}.
If not specified, \code{\link[base:tempfile]{base::tempdir()}}.}

\item{...}{additional options passed to \code{\link[base:readRDS]{base::readRDS()}}}
}
\value{
None, global environment is modified
}
\description{
Load multiple RDS files into current environment
}
\section{DEPRECATED}{
Obsoleted by \code{\link[drake:loadd]{drake::loadd()}}
}

\examples{
# Saves/loads files "mtcars.rds" and "letters.rds" in tmpdir
my_mtcars <- mtcars; my_letters <- letters
save_all(c("my_mtcars", "my_letters"))
rm(my_mtcars, my_letters)
read_all(c("my_mtcars", "my_letters"))
}
\seealso{
Other saving/loading data: 
\code{\link{save_all}()}
}
\author{
Sean Ho \href{mailto:anchor@seanho.com}{anchor@seanho.com}
}
\concept{saving/loading data}
