% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all.r
\name{save_all}
\alias{save_all}
\title{Save multiple objects as RDS files}
\usage{
save_all(.list, ...)
}
\arguments{
\item{.list}{character vector of names of objects in environment}

\item{...}{path components relative to project root}
}
\value{
None, global environment is modified
}
\description{
Save multiple objects as RDS files
}
\examples{
# Creates files "data/rds/mtcars.rds" and "data/rds/letters.rds"
# relative to project root
save_all(c("mtcars", "letters"), "data", "rds")
}
\seealso{
Other Saving/loading data: \code{\link{read_all}}
}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{Saving/loading data}
