% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all.R
\name{save_all}
\alias{save_all}
\title{Save multiple objects as RDS files}
\usage{
save_all(.list, rdsdir = NULL, ...)
}
\arguments{
\item{.list}{character vector of names of objects in environment}

\item{rdsdir}{path to directory, passed to \code{\link[base:file.path]{base::file.path()}}.
If not specified, \code{\link[base:tempfile]{base::tempdir()}}.}

\item{...}{additional options passed to \code{\link[base:readRDS]{base::saveRDS()}}}
}
\value{
None, global environment is modified
}
\description{
Save multiple objects as RDS files
}
\section{DEPRECATED}{
Obsoleted by drake caching
}

\examples{
# Saves/loads files "mtcars.rds" and "letters.rds" in tmpdir
my_mtcars <- mtcars; my_letters <- letters
save_all(c("my_mtcars", "my_letters"))
rm(my_mtcars, my_letters)
read_all(c("my_mtcars", "my_letters"))
}
\seealso{
Other saving/loading data: 
\code{\link{read_all}()}
}
\concept{saving/loading data}
