% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{VR12_items}
\alias{VR12_items}
\title{List of items in VR-12 (Veterans RAND 12-item) survey}
\format{
A character vector of 12 names.
}
\usage{
VR12_items
}
\description{
A vector containing the expected sequence of VR12 items,
as used by VR12_scores().
}
\keyword{datasets}
