% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VR12.r
\name{VR12_score}
\alias{VR12_score}
\title{Compute VR12 subscales}
\usage{
VR12_score(.data, coefs = VR12_coefs$Phone)
}
\arguments{
\item{.data}{tibble}

\item{coefs}{scoring coefficients}
}
\value{
tibble with new columns VR12PCS and VR12MCS
}
\description{
Given a tibble of VR12 item scores, append two new columns:
VR12PCS (physical) and VR12MCS (mental).
}
\details{
The input data should have at least the following integer columns:
GH1, PF2, PF4, RP2, RP3, RE2, RE3, BP2, MH3, VT2, MH4, SF2.
It may have other columns as well.
A temporary column named CONS is used; if a column of this name exists
already, it will be deleted.
If there is any missing data, imputation should be performed prior to this.

The default set of scoring coefficients is from the Interviewer form.
}
\examples{
VR12_score(data.frame(
  GH1 = 2:3, PF2 = 0:1, PF4 = 0:1, RP2 = 2:3, RP3 = 2:3, RE2 = 2:3, RE3 = 2:3,
  BP2 = 2:3, MH3 = 2:3, VT2 = 2:3, MH4 = 0:1, SF2 = 0:1))

}
\author{
Sean Ho <anchor@seanho.com>
}
\concept{VR12}
