% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_names.r
\name{which_names}
\alias{which_names}
\title{Find indices of TRUE entries in an array}
\usage{
which_names(.a, dimnames = NULL, ...)
}
\arguments{
\item{.a}{array}

\item{dimnames}{list of character vectors.
E.g., if \code{.a} is a matrix, then \code{length(dimnames) == 2},
\code{dimnames[[1]]} is a vector of row names, and
\code{dimnames[[2]]} is a vector of column names.}

\item{...}{Additional options passed to \code{\link[base:which]{base::which()}}.}
}
\value{
a data frame of named indices
}
\description{
\code{\link[base:which]{base::which()}} returns numeric indices of TRUE entries in an array.
This wrapper uses \code{dimnames} to return pairs of row/column names.
}
\details{
This was originally created to examine correlation matrices for collinearity.
Then I discovered \code{\link[corrr:stretch]{corrr::stretch()}} and that was much better.
}
\examples{
which_names(mtcars[1:6, c("vs", "am")] == 1)
which_names(matrix(1:20, nrow = 4) \%\% 3 == 0)
}
